refactor(embeddings): split provider implementations
